renamed util files and formatted Build
