+++ tensorflow/c/eager/mnist_gradients_testutil.h
